+++ freqtrade/rpc/api_server2/api_v1.py
@@ -1,9 +1,10 @@
+from freqtrade.rpc import RPC
 from fastapi import APIRouter, Depends
 
 from freqtrade import __version__
 
+from .api_models import Balances, Ping, StatusMsg, Version
 from .deps import get_config, get_rpc
-from .models import Balances, Ping, Version
 
 
 # Public API, requires no auth.
@@ -18,13 +19,21 @@ def ping():
     return {"status": "pong"}
 
 
-@router.get('/version', response_model=Version)
+@router.get('/version', response_model=Version, tags=['info'])
 def version():
     return {"version": __version__}
 
 
-@router.get('/balance', response_model=Balances)
-def balance(rpc=Depends(get_rpc), config=Depends(get_config)):
+@router.get('/balance', response_model=Balances, tags=['info'])
+def balance(rpc: RPC = Depends(get_rpc), config=Depends(get_config)):
     return rpc._rpc_balance(config['stake_currency'], config.get('fiat_display_currency', ''),)
 
 
+@router.post('/start', response_model=StatusMsg, tags=['botcontrol'])
+def start(rpc: RPC = Depends(get_rpc)):
+    return rpc._rpc_start()
+
+
+@router.post('/stop', response_model=StatusMsg, tags=['botcontrol'])
+def stop(rpc: RPC = Depends(get_rpc)):
+    return rpc._rpc_stop()
