+++ freqtrade/rpc/api_server2/uvicorn_threaded.py
@@ -1,6 +1,7 @@
 import contextlib
-import time
 import threading
+import time
+
 import uvicorn
